# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserSignUpResponse": {
				"properties": {
					"email": {
						"type": "string",
						"description": "가입된 사용자 이메일"
					},
					"name": {
						"type": "string",
						"description": "가입된 사용자 이름"
					},
					"preferCategory": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "사용자가 선택한 선호 카테고리 이름 목록"
					}
				},
				"required": [
					"email",
					"name",
					"preferCategory"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserSignUpResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/UserSignUpResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"FAILED"
						],
						"nullable": false
					},
					"error": {
						"properties": {
							"data": {
								"nullable": true
							},
							"message": {
								"type": "string",
								"nullable": true
							},
							"errorCode": {
								"type": "string",
								"nullable": true
							}
						},
						"required": [
							"data",
							"message",
							"errorCode"
						],
						"type": "object"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSignUpRequest": {
				"properties": {
					"email": {
						"type": "string",
						"description": "사용자 이메일"
					},
					"name": {
						"type": "string",
						"description": "사용자 이름"
					},
					"gender": {
						"type": "string",
						"description": "사용자 성별"
					},
					"birth": {
						"type": "string",
						"description": "생년월일"
					},
					"address": {
						"type": "string",
						"description": "기본 주소"
					},
					"detailAddress": {
						"type": "string",
						"description": "상세 주소"
					},
					"phoneNumber": {
						"type": "string",
						"description": "전화번호"
					},
					"preferences": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array",
						"description": "선호 음식 카테고리 ID 목록"
					}
				},
				"required": [
					"email",
					"name",
					"gender",
					"birth",
					"phoneNumber",
					"preferences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StoreCreateResponse": {
				"properties": {
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "생성된 가게 ID"
					},
					"name": {
						"type": "string",
						"description": "가게 이름"
					},
					"address": {
						"type": "string",
						"description": "가게 주소"
					},
					"regionId": {
						"type": "number",
						"format": "double",
						"description": "가게가 속한 지역 ID"
					},
					"regionName": {
						"type": "string",
						"description": "가게가 속한 지역 이름"
					}
				},
				"required": [
					"storeId",
					"name",
					"address",
					"regionId",
					"regionName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse__message-string--data-StoreCreateResponse__": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"properties": {
							"data": {
								"$ref": "#/components/schemas/StoreCreateResponse"
							},
							"message": {
								"type": "string"
							}
						},
						"required": [
							"data",
							"message"
						],
						"type": "object"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StoreCreateRequest": {
				"properties": {
					"name": {
						"type": "string",
						"description": "가게 이름"
					},
					"address": {
						"type": "string",
						"description": "가게 주소"
					},
					"score": {
						"type": "number",
						"format": "double",
						"description": "가게 평점"
					}
				},
				"required": [
					"name",
					"address"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewCreateResponse": {
				"properties": {
					"reviewId": {
						"type": "number",
						"format": "double",
						"description": "생성된 리뷰 ID"
					},
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "리뷰가 작성된 가게 ID"
					},
					"storeName": {
						"type": "string",
						"description": "리뷰가 작성된 가게 이름"
					},
					"userId": {
						"type": "number",
						"format": "double",
						"description": "리뷰 작성자 ID"
					},
					"nickname": {
						"type": "string",
						"description": "리뷰 작성자 닉네임"
					},
					"rating": {
						"type": "number",
						"format": "double",
						"description": "리뷰 별점"
					},
					"content": {
						"type": "string",
						"description": "리뷰 내용"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "리뷰 작성 일시"
					}
				},
				"required": [
					"reviewId",
					"storeId",
					"storeName",
					"userId",
					"nickname",
					"rating",
					"content",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse__message-string--data-ReviewCreateResponse__": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"properties": {
							"data": {
								"$ref": "#/components/schemas/ReviewCreateResponse"
							},
							"message": {
								"type": "string"
							}
						},
						"required": [
							"data",
							"message"
						],
						"type": "object"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewCreateRequest": {
				"properties": {
					"userId": {
						"type": "number",
						"format": "double",
						"description": "리뷰를 작성하는 사용자 ID"
					},
					"rating": {
						"type": "number",
						"format": "double",
						"description": "리뷰 별점"
					},
					"content": {
						"type": "string",
						"description": "리뷰 내용"
					}
				},
				"required": [
					"userId",
					"rating",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewListItem": {
				"properties": {
					"reviewId": {
						"type": "number",
						"format": "double",
						"description": "리뷰 ID"
					},
					"nickname": {
						"type": "string",
						"description": "리뷰 작성자 닉네임"
					},
					"rating": {
						"type": "number",
						"format": "double",
						"description": "리뷰 별점"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "리뷰 작성 일시"
					},
					"content": {
						"type": "string",
						"description": "리뷰 내용"
					}
				},
				"required": [
					"reviewId",
					"nickname",
					"rating",
					"createdAt",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewListResponse": {
				"properties": {
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "가게 ID"
					},
					"storeName": {
						"type": "string",
						"description": "가게 이름"
					},
					"reviews": {
						"items": {
							"$ref": "#/components/schemas/ReviewListItem"
						},
						"type": "array",
						"description": "리뷰 목록"
					}
				},
				"required": [
					"storeId",
					"storeName",
					"reviews"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse__message-string--data-ReviewListResponse__": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"properties": {
							"data": {
								"$ref": "#/components/schemas/ReviewListResponse"
							},
							"message": {
								"type": "string"
							}
						},
						"required": [
							"data",
							"message"
						],
						"type": "object"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MyReviewListItem": {
				"properties": {
					"reviewId": {
						"type": "number",
						"format": "double",
						"description": "리뷰 ID"
					},
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "가게 ID"
					},
					"storeName": {
						"type": "string",
						"description": "가게 이름"
					},
					"rating": {
						"type": "number",
						"format": "double",
						"description": "리뷰 별점"
					},
					"content": {
						"type": "string",
						"description": "리뷰 내용"
					},
					"createdAt": {
						"type": "string",
						"description": "리뷰 작성일"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "리뷰 이미지 URL 목록"
					}
				},
				"required": [
					"reviewId",
					"storeId",
					"storeName",
					"rating",
					"content",
					"createdAt",
					"images"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MyReviewListResponse": {
				"properties": {
					"reviews": {
						"items": {
							"$ref": "#/components/schemas/MyReviewListItem"
						},
						"type": "array",
						"description": "내가 작성한 리뷰 목록"
					},
					"pagination": {
						"properties": {
							"hasNext": {
								"type": "boolean",
								"description": "다음 페이지 존재 여부"
							},
							"cursor": {
								"type": "number",
								"format": "double",
								"description": "다음 조회에 사용할 cursor"
							}
						},
						"required": [
							"hasNext",
							"cursor"
						],
						"type": "object",
						"description": "페이지네이션 정보"
					}
				},
				"required": [
					"reviews",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse__message-string--data-MyReviewListResponse__": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"properties": {
							"data": {
								"$ref": "#/components/schemas/MyReviewListResponse"
							},
							"message": {
								"type": "string"
							}
						},
						"required": [
							"data",
							"message"
						],
						"type": "object"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "",
		"contact": {}
	},
	"paths": {
		"/users/signup": {
			"post": {
				"operationId": "HandleUserSignUp",
				"responses": {
					"200": {
						"description": "회원가입 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserSignUpResponse_"
								}
							}
						}
					},
					"400": {
						"description": "회원가입 요청 값이 올바르지 않습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "선호 카테고리를 찾을 수 없습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "회원가입 처리 중 서버 오류가 발생했습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "사용자 회원가입을 진행합니다.\n\n이메일, 이름, 성별, 생년월일, 전화번호, 선호 카테고리 ID 목록을 body로 전달합니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserSignUpRequest"
							}
						}
					}
				}
			}
		},
...
		"/users/mypage": {
			"get": {
				"operationId": "HandleMypage",
				"responses": {
					"200": {
						"description": "마이페이지 조회 성공",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"401": {
						"description": "로그인이 필요합니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "로그인 쿠키가 있는 사용자만 접근할 수 있는 마이페이지를 반환합니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			}
		},
...
		"/regions/{regionId}/stores": {
			"post": {
				"operationId": "HandleCreateStore",
				"responses": {
					"201": {
						"description": "가게 생성 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__message-string--data-StoreCreateResponse__"
								}
							}
						}
					},
					"400": {
						"description": "가게 생성 요청 값이 올바르지 않습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "지역을 찾을 수 없습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "가게 생성 처리 중 서버 오류가 발생했습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "특정 지역에 새로운 가게를 등록합니다.\n\npath parameter로 지역 ID를 받고, body로 가게 이름과 주소를 전달합니다.",
				"tags": [
					"Stores"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "regionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/StoreCreateRequest"
							}
						}
					}
				}
			}
		},
		"/stores/{storeId}/reviews": {
			"post": {
				"operationId": "HandleCreateReview",
				"responses": {
					"201": {
						"description": "리뷰 작성 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__message-string--data-ReviewCreateResponse__"
								}
							}
						}
					},
					"400": {
						"description": "리뷰 작성 요청 값이 올바르지 않습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "가게 또는 사용자를 찾을 수 없습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "리뷰 작성 처리 중 서버 오류가 발생했습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "특정 가게에 리뷰를 작성합니다.\n\npath parameter로 가게 ID를 받고, body로 사용자 ID, 별점, 리뷰 내용을 전달합니다.",
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReviewCreateRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "HandleListStoreReviews",
				"responses": {
					"200": {
						"description": "리뷰 목록 조회 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__message-string--data-ReviewListResponse__"
								}
							}
						}
					},
					"400": {
						"description": "가게 ID가 올바르지 않습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "가게를 찾을 수 없습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "특정 가게에 작성된 리뷰 목록을 조회합니다.\n\npath parameter로 가게 ID를 전달합니다.",
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/{userId}/reviews": {
			"get": {
				"operationId": "HandleListMyReviews",
				"responses": {
					"200": {
						"description": "내가 작성한 리뷰 목록 조회 성공",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__message-string--data-MyReviewListResponse__"
								}
							}
						}
					},
					"400": {
						"description": "사용자 ID 또는 cursor가 올바르지 않습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "사용자를 찾을 수 없습니다.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "특정 사용자가 작성한 리뷰 목록을 cursor 기반으로 조회합니다.\n\npath parameter로 사용자 ID를 받고, query parameter로 cursor를 전달할 수 있습니다.",
				"tags": [
					"Reviews"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}